- Renamed AbstractDataAnalysis to AbstractCharacterisation and improved interface - Renamed DataAnalysis to EDNACharacterisation - Removed duplicate dc_from_edna_output and moved it to EDNACharacterisation - Made get_default_characterisation_parameters member of EDNACharacterisation
