+++ HardwareObjects/ALBA/ALBADataAnalysis.py
@@ -1,7 +1,7 @@
 from xaloc import XalocJob
 from XSDataCommon import XSDataFile, XSDataString
 from XSDataMXCuBEv1_3 import XSDataResultMXCuBE
-from HardwareRepository.HardwareObjects.DataAnalysis import DataAnalysis
+from HardwareRepository.HardwareObjects.EDNACharacterisation import EDNACharacterisation
 from PyTango import DeviceProxy
 import os
 import time
@@ -17,9 +17,9 @@ root = os.environ["POST_PROCESSING_SCRIPTS_ROOT"]
 sls_script = os.path.join(root, "edna-mx/strategy/edna-mx.strategy.sl")
 
 
-class ALBADataAnalysis(DataAnalysis):
+class ALBADataAnalysis(EDNACharacterisation):
     def init(self):
-        DataAnalysis.init(self)
+        EDNACharacterisation.init(self)
 
     def prepare_edna_input(self, edna_input):
