+++ configuration/desy_p11/data-analysis.xml
@@ -1,4 +1,4 @@
-<object class="DataAnalysis">
+<object class="EDNACharacterisation">
   <!-- Path to script that launches edna -->
   <edna_command>PATH_TO_EDNA_LAUNCH_SCRIPT</edna_command>
   <edna_default_file>edna_defaults.xml</edna_default_file>
