+++ configuration/esrf_id29/data-analysis.xml
@@ -1,4 +1,4 @@
-<object class="DataAnalysis">
+<object class="EDNACharacterisation">
   <edna_command>/users/blissadm/local/bin/edna_script.sh</edna_command>
   <edna_default_file>/users/blissadm/local/HardwareRepository/edna_defaults.xml</edna_default_file>
   <channel type="spec" version="lid292:exp" name="beam_size_x">MXSPEC_PARS["beamsize"]["x"]</channel>
