+++ configuration/soleil_px1/data-analysis.xml
@@ -1,4 +1,4 @@
-<object class="DataAnalysis">
+<object class="EDNACharacterisation">
   <edna_command>/users/blissadm/local/bin/edna_script_bes.py</edna_command>
   <edna_default_file>edna_defaults.xml</edna_default_file>
 </object>
